# [instrument] Improve sampled timer and stopwatch performance